Add files via upload
Developed a web-based Sudoku Solver application that allows users to solve Sudoku puzzles effortlessly. The application uses HTML, JavaScript and employs backtracking algorithms to find the optimal solution for any given puzzle. The main components of the project are: A 9x9 board. A ‘Solve’ button that will trigger the algorithm and solve the puzzle. A ‘Clear’ button to clear previous input.